Add answer controller

+++ Api/internal/domains/models/answer.go
@@ -1,7 +1,7 @@
 package models
 
 type Answer struct {
-	Id      int    `db:"answer_id" json:"id"`
+	Id      int64  `db:"answer_id" json:"id"`
 	Text    string `json:"text"`
 	IsRight bool   `json:"isRight"`
 }
