+++ Api/internal/domains/models/question.go
@@ -1,9 +1,9 @@
 package models
 
 type Question struct {
-	Id      int    `db:"question_id" json:"id"`
+	Id      int64  `db:"question_id" json:"id"`
 	Text    string `json:"text,omitempty"`
 	Image   []byte `json:"image,omitempty"`
-	ScoreId int    `db:"score_id" json:"scoreId"`
-	TypeId  int    `db:"type_id" json:"typeId"`
+	ScoreId int64  `db:"score_id" json:"scoreId"`
+	TypeId  int64  `db:"type_id" json:"typeId"`
 }
