+++ Api/internal/domains/models/question_type.go
@@ -1,6 +1,6 @@
 package models
 
 type QuestionType struct {
-	Id   int    `db:"type_id" json:"id"`
+	Id   int64  `db:"type_id" json:"id"`
 	Name string `json:"name"`
 }
